+++ history.md
@@ -1,7 +1,11 @@
 ### ToDo
 - ButtonView的uploader导入功能考虑实现逻辑
 - ListView的整体逻辑实现
+- ListView的TableMenu实现级联
 - SearchView
+
+### 重要
+- SimpleTableView的插槽和菜单逻辑同步
 
 ### 4.0.1
 - 基于v4版本的ant-design和complex-component构建模板